Refactor and adeded a model class
