+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:theme="@style/Theme.Watch"
         tools:targetApi="31">
         <activity
-            android:name=".WatchActivity"
+            android:name=".activities.WatchActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
