+++ app/src/main/java/ie/setu/watch/models/WatchModel.kt
@@ -1,0 +1,3 @@
+package ie.setu.watch.models
+
+data class WatchModel(var title: String = "")
