+++ app/src/main/res/layout/activity_watch.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ie.setu.watch.WatchActivity">
+    tools:context="ie.setu.watch.activities.WatchActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
